+++ grid2op/Environment/multiMixEnv.py
@@ -13,7 +13,7 @@ import copy
 from typing import Any, Dict, Tuple, Union, List
 
 from grid2op.dtypes import dt_int, dt_float
-from grid2op.Space import GridObjects, RandomObject
+from grid2op.Space import GridObjects, RandomObject, DEFAULT_N_BUSBAR_PER_SUB
 from grid2op.Exceptions import EnvError, Grid2OpException
 from grid2op.Observation import BaseObservation
 
@@ -161,6 +161,7 @@ class MultiMixEnvironment(GridObjects, RandomObject):
         envs_dir,
         logger=None,
         experimental_read_from_local_dir=False,
+        n_busbar=DEFAULT_N_BUSBAR_PER_SUB,
         _add_to_name="",  # internal, for test only, do not use !
         _compat_glop_version=None,  # internal, for test only, do not use !
         _test=False,
@@ -217,6 +218,7 @@ class MultiMixEnvironment(GridObjects, RandomObject):
                         backend=bk,
                         _add_to_name=_add_to_name,
                         _compat_glop_version=_compat_glop_version,
+                        n_busbar=n_busbar,
                         test=_test,
                         logger=this_logger,
                         experimental_read_from_local_dir=experimental_read_from_local_dir,
@@ -225,6 +227,7 @@ class MultiMixEnvironment(GridObjects, RandomObject):
                 else:
                     env = make(
                         env_path,
+                        n_busbar=n_busbar,
                         _add_to_name=_add_to_name,
                         _compat_glop_version=_compat_glop_version,
                         test=_test,