+++ grid2op/MakeEnv/Make.py
@@ -247,6 +247,7 @@ def _aux_make_multimix(
     dataset_path,
     test=False,
     experimental_read_from_local_dir=False,
+    n_busbar=2,
     _add_to_name="",
     _compat_glop_version=None,
     logger=None,
@@ -258,6 +259,7 @@ def _aux_make_multimix(
     return MultiMixEnvironment(
         dataset_path,
         experimental_read_from_local_dir=experimental_read_from_local_dir,
+        n_busbar=n_busbar,
         _test=test,
         _add_to_name=_add_to_name,
         _compat_glop_version=_compat_glop_version,
@@ -272,6 +274,7 @@ def make(
     test : bool=False,
     logger: Optional[logging.Logger]=None,
     experimental_read_from_local_dir : bool=False,
+    n_busbar=2,
     _add_to_name : str="",
     _compat_glop_version : Optional[str]=None,
     **kwargs
@@ -307,6 +310,9 @@ def make(
         it might not be easily usable. If you encounter issues with pickle or multi
         processing, you can set this flag to ``True``. See the doc of 
         :func:`grid2op.Environment.BaseEnv.generate_classes` for more information.
+        
+    n_busbar: ``int``
+        Number of independant busbars allowed per substations. By default it's 2.
         
     kwargs:
         Other keyword argument to give more control on the environment you are creating. See
@@ -402,6 +408,7 @@ def make(
             dataset_path=dataset,
             _add_to_name=_add_to_name_tmp,
             _compat_glop_version=_compat_glop_version_tmp,
+            n_busbar=n_busbar,
             **kwargs
         )
 
@@ -441,6 +448,7 @@ def make(
         return make_from_path_fn(
             dataset_path=ds_path,
             logger=logger,
+            n_busbar=n_busbar,
             _add_to_name=_add_to_name,
             _compat_glop_version=_compat_glop_version,
             experimental_read_from_local_dir=experimental_read_from_local_dir,
@@ -454,6 +462,7 @@ def make(
         return make_from_path_fn(
             real_ds_path,
             logger=logger,
+            n_busbar=n_busbar,
             experimental_read_from_local_dir=experimental_read_from_local_dir,
             **kwargs
         )
@@ -472,6 +481,7 @@ def make(
     return make_from_path_fn(
         dataset_path=real_ds_path,
         logger=logger,
+        n_busbar=n_busbar,
         experimental_read_from_local_dir=experimental_read_from_local_dir,
         **kwargs
     )